Restructure app to remove circular dependancy
The import of app in models and models in app creates a circular
dependancy that makes it hard to test/develop the database independantly
of the rest of the app. This removes the global reference to app from
most places, uses a blueprint instead of app in the views, uses an app
factory in the __init__.py instead of just defining app and allows apps
to be created with different configuration

Normally static and templates would have been moved into app as this is
where flask expects them, but I instead se their location when
configuring the app in create app, to not break the static website and
to not have to change the directory structure too much
