+++ app.py
@@ -1,99 +1,3 @@
-from flask import Flask, render_template, redirect, url_for, request, flash
-from chengyu import select
-from werkzeug.security import generate_password_hash
-from os import listdir, environ
-from time import time
-from json import loads
-from flask_login import LoginManager, current_user, login_user, logout_user
-from forms import Registration, Login
-from sqlalchemy.exc import IntegrityError
-from models import db, User, Game, Play
+from app import app, db
 
-app = Flask(__name__)
-app.config['SECRET_KEY'] = 'a259223f6fbaa6b4678936fa'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///app.db'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-
-
-login = LoginManager(app)
-db.init_app(app)
-
-with app.app_context():
-    db.create_all()
-    db.session.commit()
-
-@login.user_loader
-def load_user(id):
-    return User.query.get(int(id))
-
-def getPuzzle(chengyu):
-    return {
-        "options": "".join([c["chinese"] for c in chengyu]),
-        "answer": chengyu[0]["chinese"],
-        "question": chengyu[0]["english"]
-    }
-
-@app.route("/chengyu")
-def chengyu():
-    return getPuzzle(select('static/chengyu.json'))
-
-@app.route("/")
-@app.route("/index")
-def daily():
-    key = bytes(int(time())//(60*60*24))
-    return render_template('index.html',
-        title="Daily Puzzle",
-        puzzle=getPuzzle(select('static/chengyu.json', 4, key))
-    )
-
-@app.route("/random")
-def random():
-    return render_template('index.html',
-        title="Random Puzzle",
-        puzzle=getPuzzle(select('static/chengyu.json', 4))
-    )
-
-@app.route("/history")
-def history():
-    with open("history.json") as file: games = loads(file.read())
-    return render_template("history.html", games=games)
-
-@app.route("/registration", methods=['GET', 'POST'])
-def register():
-    form = Registration()
-    if request.method == "POST":
-        if form.validate_on_submit():
-            user = User(
-                username=form.username.data,
-                email=form.email.data,
-                password=generate_password_hash(form.password.data))
-            db.session.add(user)
-            try:
-                db.session.commit()
-                login_user(user, remember=True) # TODO make this optional
-                return redirect(url_for('daily'))
-            except IntegrityError as error:
-                flash(error, 'error')
-                return render_template(
-                    'register.html',
-                    form=form,
-                    title="Registration Page",
-                )
-    else: return render_template('register.html', form=form, title="Registration Page")
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    form = Login()
-    if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
-        if user is None or not user.checkPassword(form.password.data):
-            flash('Invalid username or password', 'error')
-            return redirect(url_for('login'))
-        login_user(user)
-        return redirect(url_for('daily'))
-    return render_template('login.html', title='Sign In', form=form)
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for("daily"))
+db.create_all()
